+++ src/shared/ui/calendar/index.ts
@@ -2,7 +2,7 @@ export { default as CalendarCell } from './ui/calendar-cell.tsx'
 export { default as Calendar } from './ui/calendar'
 export { default as HeaderButton } from './ui/header-button'
 export { default as HeaderContainer } from './ui/header-container'
-export { default as HearderMonthLabel } from './ui/header-month-label'
+export { default as HeaderMonthLabel } from './ui/header-month-label'
 export { default as InfiniteCalendar } from './ui/infinite-calendar'
 export { default as MonthlyCalendar } from './ui/montly-calendar'
 export { default as PrivateScheduleBadge } from './ui/private-schedule-badge'
@@ -10,7 +10,7 @@ export { default as PrivateScheduleModal } from './ui/private-schedule-modal'
 export { default as AddScheduleButton } from './ui/add-schedule-button'
 export { default as ScheduleBadge } from './ui/schedule-badge'
 export { default as AppointmentBadge } from './ui/appointment-badge'
-export { default as RenderScheduleBadges } from './ui/render-schdule-badges'
+export { default as RenderScheduleBadges } from './ui/render-schedule-badges'
 
 export * from './utils/date'
 export * from './utils/get-day-info'
