+++ src/shared/ui/calendar/ui/render-schedule-badges.tsx
